Move AppContext switch test to separate file
