+++ test/System.IdentityModel.Tokens.Jwt.Tests/JwtSecurityTokenHandlerNonParallelRunTests.cs
@@ -1,0 +1,9 @@
+﻿using Xunit;
+
+namespace System.IdentityModel.Tokens.Jwt.Tests
+{
+    [CollectionDefinition(nameof(JwtSecurityTokenHandlerNonParallelRunTests), DisableParallelization = true)]
+    public class JwtSecurityTokenHandlerNonParallelRunTests
+    {
+    }
+}
