+++ test/System.IdentityModel.Tokens.Jwt.Tests/JwtSecurityTokenHandlerTests.cs
@@ -18,7 +18,6 @@ using TokenLogMessages = Microsoft.IdentityModel.Tokens.LogMessages;
 
 namespace System.IdentityModel.Tokens.Jwt.Tests
 {
-    [CollectionDefinition("JwtSecurityTokenHandlerTests", DisableParallelization = true)]
     public class JwtSecurityTokenHandlerTests
     {
 
@@ -54,108 +53,6 @@ namespace System.IdentityModel.Tokens.Jwt.Tests
 
             foreach (var value in claimValues)
                 Assert.NotNull(aTypeClaims.SingleOrDefault(c => c.Value == value));
-        }
-
-        [Fact]
-        
-        public void JwtSecurityTokenHandler_CreateToken_AddShortFormMappingForRsaOAEPEnabled()
-        {
-            AppContext.SetSwitch(X509EncryptingCredentials._useShortNameForRsaOaepKey, true);
-            var encryptingCredentials = new X509EncryptingCredentials(Default.Certificate);
-            SecurityTokenDescriptor tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Issuer = Default.Issuer,
-                IssuedAt = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Subject = new ClaimsIdentity(Default.PayloadJsonClaims),
-                NotBefore = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Expires = DateTime.UtcNow.Subtract(new TimeSpan(0, 10, 0)),
-                SigningCredentials = Default.AsymmetricSigningCredentials,
-                EncryptingCredentials = encryptingCredentials,
-                TokenType = "JWE"
-            };
-
-            JwtSecurityTokenHandler tokenHandler = new JwtSecurityTokenHandler();
-            JwtSecurityToken token = tokenHandler.CreateJwtSecurityToken(tokenDescriptor);
-
-            Assert.NotNull(token);
-            Assert.Equal(token.Header.Alg, SecurityAlgorithms.RsaOAEP);
-        }
-
-        [Fact]
-        public void JwtSecurityTokenHandler_CreateToken_AddShortFormMappingForRsaOAEPDisabled()
-        {
-            AppContext.SetSwitch(X509EncryptingCredentials._useShortNameForRsaOaepKey, false);
-            var encryptingCredentials = new X509EncryptingCredentials(Default.Certificate);
-            SecurityTokenDescriptor tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Issuer = Default.Issuer,
-                IssuedAt = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Subject = new ClaimsIdentity(Default.PayloadJsonClaims),
-                NotBefore = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Expires = DateTime.UtcNow.Subtract(new TimeSpan(0, 10, 0)),
-                SigningCredentials = Default.AsymmetricSigningCredentials,
-                EncryptingCredentials = encryptingCredentials,
-                TokenType = "JWE"
-            };
-
-            JwtSecurityTokenHandler tokenHandler = new JwtSecurityTokenHandler();
-            JwtSecurityToken token = tokenHandler.CreateJwtSecurityToken(tokenDescriptor);
-
-            Assert.NotNull(token);
-            Assert.Equal(token.Header.Alg, SecurityAlgorithms.RsaOAEP);
-        }
-
-        [Fact]
-        public void JsonWebTokenHandler_CreateToken_AddShortFormMappingForRsaOAEPEnabled()
-        {
-            AppContext.SetSwitch(X509EncryptingCredentials._useShortNameForRsaOaepKey, true);
-            var encryptingCredentials = new X509EncryptingCredentials(Default.Certificate);
-            SecurityTokenDescriptor tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Issuer = Default.Issuer,
-                IssuedAt = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Subject = new ClaimsIdentity(Default.PayloadJsonClaims),
-                NotBefore = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Expires = DateTime.UtcNow.Subtract(new TimeSpan(0, 10, 0)),
-                SigningCredentials = Default.AsymmetricSigningCredentials,
-                EncryptingCredentials = encryptingCredentials,
-                TokenType = "JWE"
-            };
-
-            JsonWebTokenHandler tokenHandler = new JsonWebTokenHandler();
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            JsonWebToken jsonToken = tokenHandler.ReadToken(token) as JsonWebToken;
-
-            Assert.NotNull(jsonToken);
-            Assert.NotEqual(jsonToken._alg, SecurityAlgorithms.RsaOaepKeyWrap);
-            Assert.Equal(jsonToken._alg, SecurityAlgorithms.RsaOAEP);
-        }
-
-
-        [Fact]
-        public void JsonWebTokenHandler_CreateToken_AddShortFormMappingForRsaOAEPDisabled()
-        {
-            AppContext.SetSwitch(X509EncryptingCredentials._useShortNameForRsaOaepKey, false);
-            var encryptingCredentials = new X509EncryptingCredentials(Default.Certificate);
-            SecurityTokenDescriptor tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Issuer = Default.Issuer,
-                IssuedAt = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Subject = new ClaimsIdentity(Default.PayloadJsonClaims),
-                NotBefore = DateTime.UtcNow.Subtract(new TimeSpan(1, 0, 0)),
-                Expires = DateTime.UtcNow.Subtract(new TimeSpan(0, 10, 0)),
-                SigningCredentials = Default.AsymmetricSigningCredentials,
-                EncryptingCredentials = encryptingCredentials,
-                TokenType = "JWE"
-            };
-
-            JsonWebTokenHandler tokenHandler = new JsonWebTokenHandler();
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            JsonWebToken jsonToken = tokenHandler.ReadToken(token) as JsonWebToken;
-
-            Assert.NotNull(jsonToken);
-            Assert.Equal(jsonToken._alg, SecurityAlgorithms.RsaOaepKeyWrap);
-            Assert.NotEqual(jsonToken._alg, SecurityAlgorithms.RsaOAEP);
         }
 
         [Theory, MemberData(nameof(CreateJWEWithPayloadStringTheoryData))]
